--- src/components/Icon.Types.ts
@@ -1,8 +1,0 @@
-export type IconPath = object | string
-export type IconProps = {
-  iconPath: IconPath
-  viewBox?: string
-  gId?: string
-  size?: number
-  rtlFlip?: boolean
-}
